Refactor Tables P1
